Move Validator class to helper directory
