+++ server/src/controllers/membershipController.js
@@ -1,5 +1,5 @@
 import db from '../models/index';
-import Validator from './validator';
+import Validator from '../helpers/validator';
 
 let errorMessage;
